Create CoinviewRule base class
